Fixing saving. Fixed camera flow. Forgot to set "CameraGrabber.waiting=True" for several exposure time changed calls so camera was not restarting. Fixed error when trying to start already running in_stream after running hysteresis loop tool CameraGrabber now adds uint16 frames to buffer and 32bits only used during calculation. Saved as uint16s also.

+++ SweeperUIs.py
@@ -170,8 +170,8 @@ class AnalyserSweepDialog(QDialog):
                 frame = self.camera_grabber.snap()
             cv2.imshow(self.parent.stream_window,
                        (self.parent.frame_processor._process_frame(
-                           frame.astype(np.int32)
-                       ).astype(np.uint16)
+                           frame
+                       )
                         ))
             if self.roi:
                 x, y, w, h = self.roi
@@ -379,8 +379,8 @@ class FieldSweepDialog(QDialog):
             self.sweep_line.setData(fields, intensities)
             cv2.imshow(self.parent.stream_window,
                        (self.parent.frame_processor._process_frame(
-                           frame.astype(np.int32)
-                       ).astype(np.uint16)
+                           frame
+                       )
                         )
                        )
             cv2.waitKey(1)
